Move `withTmpFile` to `util.ts` and update tests to use it
