+++ pr-checks/changelog/validate.test.mts
@@ -1,33 +1,15 @@
 import assert from "node:assert/strict";
-import * as fs from "node:fs";
-import * as os from "node:os";
-import * as path from "node:path";
 import { describe, it } from "node:test";
 
+import { withTmpFile } from "../../src/util.js";
+
 import {
+  hasValidChangenoteCategory,
   isValidChangenoteContent,
   isValidChangenoteFile,
   isValidChangenoteFilename,
-  hasValidChangenoteCategory,
   VALID_CHANGE_NOTE_CATEGORIES,
 } from "./validate.mjs";
-
-async function withTmpFile<T>(
-  baseFileName: string,
-  contents: string,
-  body: (filePath: string) => Promise<T> | T,
-): Promise<T> {
-  const tmpDir = fs.mkdtempSync(
-    path.join(os.tmpdir(), "changetool-validate-test-"),
-  );
-  try {
-    const filePath = path.join(tmpDir, baseFileName);
-    fs.writeFileSync(filePath, contents);
-    return await body(filePath);
-  } finally {
-    fs.rmSync(tmpDir, { recursive: true, force: true });
-  }
-}
 
 await describe("isValidChangenoteContent", async () => {
   await it("recognizes an unordered Markdown list", () => {
